install octokit sdk. refactor background into separate comp
