+++ app/components/background.tsx
@@ -1,0 +1,9 @@
+import "~/styles/background.css";
+
+import { HTMLAttributes, PropsWithChildren } from "react";
+
+export default function Welcome(
+  props: PropsWithChildren<HTMLAttributes<HTMLElement>>
+) {
+  return <section id="background" {...props} />;
+}
