--- app/components/projects.tsx
@@ -1,3 +1,0 @@
-export default function Projects() {
-  return <h1>Projects</h1>;
-}
