--- app/components/skills.tsx
@@ -1,3 +1,0 @@
-export default function Skills() {
-  return <h1>Skills</h1>;
-}
