+++ app/routes/projects.tsx
@@ -1,0 +1,5 @@
+import { Text } from "@mantine/core";
+
+export default function Projects() {
+  return <Text>Projects</Text>;
+}
