+++ app/styles/index.css
@@ -1,34 +1,3 @@
-@keyframes gradient {
-  0% {
-    background-position: 0% 50%;
-  }
-  50% {
-    background-position: 100% 50%;
-  }
-  100% {
-    background-position: 0% 50%;
-  }
-}
-
-#welcome {
-  background: linear-gradient(
-    160deg,
-    var(--mantine-color-yellow-3),
-    var(--mantine-color-grape-3),
-    var(--mantine-color-teal-3),
-    var(--mantine-color-pink-3),
-    var(--mantine-color-cyan-3)
-  );
-  background-size: 200% 200%;
-  height: 100vh;
-  width: 100vw;
-  animation: gradient 5s ease infinite;
-  padding: var(--mantine-spacing-xl);
-  display: flex;
-  justify-content: center;
-  position: relative;
-}
-
 .content {
   padding: var(--mantine-spacing-md);
   display: flex;
