+++ package.json
@@ -21,6 +21,7 @@
     "@remix-run/serve": "^2.6.0",
     "@tabler/icons-react": "^2.47.0",
     "isbot": "^4.1.0",
+    "octokit": "^3.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
